added dynamic s expressions and a minimal unit test for them.
The parsing is done by the previous implementation, so there is minimal
duplication of effort here.  These were added mainly to make it easier to work
with the s expressions when crafting messages to be sent over the network.
counter-intuitively, I think that this will make memory management easier  when
implementing the messaging protocol as well.