finished refactoring reader code.  Untested.
The reader code now implements the new error system and is agnostic to the type
of memory layout used.  In fact, the reader doesn't mess around with weird
casting or otherwise manage memory in any way.  It uses sexp functions to do all
that, which is how it should have been done from the start.