UNTESTED refactored sexp serialization code.
added an result type declaration for vectors.  In the future I may do a complete
refactor/rewrite of the vector library to use the error types.

Before this refactor, there was a lot of code duplication in the
printer/serialization functions.  Now, since I have embraced malloc, I have
unified the printing to the serialization functions.  The printer functions are
mostly a wrapper around the serialization functions.