fixed NULL dereferencing on NIL sexp.
NIL is implemented as a NULL pointer in the sexp interface.  Function calls were
attempting to access various fields in a NIL sexp, which generated a seg fault.
The best way to fix this would likely be to write getter/setter functions for
these fields. For now, I am just testing for NIL.